+++ src/ui/mainwindow.cpp
@@ -32,9 +32,6 @@ MainWindow::MainWindow(ConfigManager *cfManager, QWidget *parent)
   setBaseSize(size());
 
   ToolBar = new QToolBar;
-  DockWidget = new QDockWidget("Preview panel", this, this->windowFlags());
-  previewPanel = new PreviewPanel;
-  DockWidget->setWidget(previewPanel);
 
   isPintotop = 0;
 
@@ -71,6 +68,13 @@ MainWindow::MainWindow(ConfigManager *cfManager, QWidget *parent)
 
   plainTextEdit = new TextEditor(configManager);
   this->setCentralWidget(plainTextEdit);
+
+  previewPanel = new QTextBrowser(this);
+  DockWidget = new QDockWidget(tr("Preview panel"), this);
+  DockWidget->setAllowedAreas(Qt::LeftDockWidgetArea | Qt::RightDockWidgetArea);
+  DockWidget->setWidget(previewPanel);
+  this->addDockWidget(Qt::RightDockWidgetArea, DockWidget);
+  DockWidget->setVisible(0);
 
   connect(ui->actionNew, &QAction::triggered, plainTextEdit,
           &TextEditor::newDocument);
@@ -159,6 +163,8 @@ MainWindow::MainWindow(ConfigManager *cfManager, QWidget *parent)
     }
     isPintotop = !isPintotop;
   });
+  connect(ui->actionPreview_panel, &QAction::triggered,
+          [&]() { DockWidget->setVisible(!DockWidget->isVisible()); });
 
   connect(plainTextEdit, &TextEditor::changeTitle, this,
           &MainWindow::changeWindowTitle);
@@ -259,6 +265,7 @@ void MainWindow::normalMode(bool first)
     ToolBar->addAction(ui->actionQuit);
     ToolBar->addAction(ui->actionAbout);
     ToolBar->addAction(ui->actionSticky_note_mode);
+    ToolBar->addAction(ui->actionPreview_panel);
   }
   ui->actionPreferences->setEnabled(1);
   currentMode = 0;
@@ -285,4 +292,6 @@ void MainWindow::stickyNoteMode()
 void MainWindow::documentWasModified()
 {
   setWindowModified(plainTextEdit->document()->isModified());
+
+  previewPanel->setSource(plainTextEdit->currentFile);
 }