+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [UNRELEASED]
+
+### 🌵 Features
+
+- Preview panel
+
 ## 0.1.3-alpha2
 
 The main purpose of this release is to fix some bugs in the previous alpha version.